update logging framework to logrus update go to 1.15 update individual logging on each filtering event update packages a little
fix bug where stock wasn't actually updating

move getStocks to main
move filtering to designated package
