+++ internal/orobouros.go
@@ -1,6 +1,6 @@
 package internal
 
-// Ouroboros contains a ring buffer style stack that can have a Rasterized array of the current state printed out
+// Ouroboros contains a ring buffer style stack that can have a rasterized array of the current state printed out
 type Ouroboros struct {
 	Capacity  int
 	beginning int
